\name{edist.face3d}

\alias{edist.face3d}

\title{Find the Euclidean distances from a particular location for each of a set of points.}

\description{
  This function finds the Euclidean distances from a particular location for each of a set of points.
}

\usage{
edist.face3d(x, x0)
}

\arguments{
\item{x}{a matrix with three columns, where each row defines a point.}
\item{x0}{a vector of length three defining a reference position.}
}

\details{
Finds the Euclidean distances from a particular location for each of a set of points.
}

\value{
A vector containing the Euclidean distances.
}
