\name{plotpca.face3d}
\alias{plotpca.face3d}

\title{Interactive plots of three-dimensional shape data, including principal components.}

\description{
This function plots three-dimensional shape data characterised by landmarks.  The \code{rgl} package is used for display.  Individual and mean shapes can be plotted, with landmarks optionally connected by lines.  Principal components can be displayed as animations.  A random tour through the space defined by the first few principal components can be created.  Where there is more than one group of shape data present, and one group can be interpreted as a reference or control, then a 'closest control' can also be displayed.  The control of the function is largely interactive, using the \code{rpanel} and \code{tclck} packages. 
}

\usage{
plotpca.face3d(gpa, npc = 6, connect = list(),
              group = NA, control = NA, title = '3D PCA display', rgl.size = 2)
}

\arguments{
\item{gpa}{a list object, created by the \code{procGPA} function of the \code{shapes} package.}
\item{npc}{the number of principal components to be explored.}
\item{connect}{a list defining the landmarks which are to be connected by lines.  Each component of the 
	           list defines a separate set of lines, so that objects which consist of collection of lines
	           can be represented.}
\item{group}{a vector of groups indicators (numeric or character values) or a factor.}
\item{control}{a numeric or character value which defined the reference or control group, if appropriate,
	           when a grouping factor is present.}
\item{title}{a character value which will be printed as the title of the interactive control panel.}
\item{rgl.size}{the size of plotted points in \code{rgl} plots.}
}

\value{
  the name of the \code{R} object which identifies the control panel.
}

\details{
The function launches a control panel which allows the user to explore the data interactively.  The items included in the panel are described briefly below.
  
\emph{Show mean}.
This checkbox determines whether the mean shape is displayed.
  
\emph{Show data}.
This checkbox determines whether individual shapes are displayed.  Only one shape is displayed at a time.  The adjacent pair of buttons allows the user to pass through the dataset.  A standard \code{R} plot is also created to show the locations of the data points in the space defined by the first two principal component scores.  As individual shapes are displayed in the \code{rgl} window, the identifier of the observation is printed on the scores plot.
  
\emph{Control mean}.
When a \code{group} variable and a \code{control} code have both been identified, this checkbox determines whether the mean value of the control group is displayed.  This mean value is calculated by averaging the control data in the space defined by the first \code{npc} principal components and converting this to a shape.
  
\emph{Closest control}.
This checkbox determines whether the 'closest control' of a particular shape is displayed.  This applies only to data which are from groups other than the control group.  A 'normal region' in the space of the first \code{npc} principal component scores is defined by constraining the Mahalonibis distance from the control mean to lie within below the 95th percentile of a chi-squared(\code{npc}) distribution.  If a particular shape lies outside this region then the projection onto the 95% contour is identified and converted to a shape.
  
\emph{Scores plot}.
This button generates a plot of the first two principal component scores of the data.
  
\emph{Identify points}.
When this button ios pressed, the user can click on the scores plot to identify a particular data point.  The corresponding shape will be displayed in the \code{rgl} window.
  
\emph{Theta, Phi}.
These two pairs of buttons can be used to control the orientation of the displayed shapes.  (Interactive control of orientation is always available by dragging the mouse across the \code{rgl} window.)
  
\emph{Principal components}.
This set of radio buttons allows individual principal components to be identified.  The adjacent \emph{Animate} button launches an animation of the shape change associated with the identified principal component.  If the \emph{Spin animation} box is checked then the orientation of the shapes will change continuously as the animation proceeds.  The \emph{Show extremes} checkbox provides an alternative means of characterising shape change in a static manner, by displaying the extreme positions reached by the animation.  These positions are defined by a standardised distance of 3 units along the principal component axis.  The \emph{Tour} button launches an animation which moves smoothly between 5 randomly selected points in the space defined by the first \code{npc} principal component scores.  In order to enhance the  shape changes, these points are generated by normal distributions with standard deviation 1.5, rather than 1.
  
The reference below describes the methods used by the function to plot the data.
}

\references{
Bowman, A.W. and Bock, M.T. (2006).  Exploring variation in three-dimensional shape data.  \emph{J. Comp. Graph. Stat.}, 15, 524-541.
}

